requirements.txt
python_dotenv
netCDF4
Magics